Disallowing blank pixels
Although the NES has a concept of blank pixels in sprites (i.e., ones
where the background is showing through), the fact that a pixel is blank
isn't implied in the binary image format. (Rather, one color is defined
to be "blank" programmatically). So not allowing it here for now.