Improved performance of the overview
Now storing all pixels a 2 dimensional array. Each root element in the
array contains the pixels of only one sprite. This way, when one pixel
in one sprite changes, only that sprite's array changes and only that
overview gets rerendered.

Before, we were storing all pixels in all sprites as one flat array, and
so Vue was registering every pixel change as a change to the whole
array, which made it rerender every sprite in the overview. So it tended
to be slower when there were lots of sprites.